Check published version.
